Extract `IpPacket` into dedicated file
